add 3x3x6

+++ gen_reduced_compgraph_3x3x6.py
@@ -6,25 +6,9 @@ from itertools import count
 import networkx as nx
 import os
 
-#BH = 32
-#BW = 32
-#A_intcoeff = 17
-#B_intcoeff = 1
-#min_terms = 3
-#A_reduced_exprs_fn = "A_reduced_exprs_32x32.pickle"
-#A_eval_order_fn = "A_eval_order_32x32.pickle"
-#A_replacements_fn = "A_replacements_32x32.pickle"
-#B_replacements_fn = "B_replacements_32x32.pickle"
-#B_reduced_exprs_fn = "B_reduced_exprs_32x32.pickle"
-#m_txt_fn = "m_32x32.txt"
-
-BH = 3
-BW = 3
-A_intcoeff = 1
 B_intcoeff = 8
 min_terms = 2
 A_reduced_exprs_fn = "A_reduced_exprs_3x3x6.pickle"
-A_eval_order_fn = "A_eval_order_3x3x6.pickle"
 A_replacements_fn = "A_replacements_3x3x6.pickle"
 B_reduced_exprs_fn = "B_reduced_exprs_3x3x6.pickle"
 B_eval_order_fn = "B_eval_order_3x3x6.pickle"
@@ -183,9 +167,9 @@ def compute_min_intermediates_eval_order(dep_graph, inputs, outputs, tmp_prefix=
     reduced_exprs = [tmp_replaced_graph.nodes[sp.Symbol(i)]['expr'] for i in outputs]
     return tmp_name_eval_order, reduced_exprs
 
-def parse_cse_gen_assignments(expr_list, rep_filename, global_tmp_prefix='Ax', tmp_replace_prefix='Axx', fm_tmp_prefix='mA', inputs_prefix='A_'):
+def parse_cse_gen_assignments(expr_list, rep_filename, global_tmp_prefix='Bx', tmp_replace_prefix='Bxx', fm_tmp_prefix='mB', inputs_prefix='B_'):
     n_exprs = len(expr_list)
-    inputs = [f'{inputs_prefix}{i+1}_{j+1}' for i in range(BH) for j in range(BW)]
+    inputs = [f'{inputs_prefix}{i+1}_{j+1}' for i in range(3) for j in range(6)]
     outputs = [f'{fm_tmp_prefix}{i+1}' for i in range(n_exprs)]
     if os.path.isfile(rep_filename):
         with open(rep_filename, 'rb') as f:
@@ -240,14 +224,18 @@ for line in s.split("\n"):
         A_expr_str_list.append(A_expre.string)
         B_expr_str_list.append(B_expre.string)
 
-A_expr_list = [sp.parsing.sympy_parser.parse_expr(s)*A_intcoeff for s in A_expr_str_list]
+A_expr_list = [sp.parsing.sympy_parser.parse_expr(s) for s in A_expr_str_list]
 
-A_eval_order, A_reduced_exprs = parse_cse_gen_assignments(A_expr_list, A_replacements_fn, global_tmp_prefix='Ax', tmp_replace_prefix='Axx', fm_tmp_prefix='mA', inputs_prefix='A_')
+if not (os.path.isfile(A_replacements_fn) and os.path.isfile(A_reduced_exprs_fn)):
+    A_expr_list = [sp.parsing.sympy_parser.parse_expr(s) for s in A_expr_str_list]
 
-with open(A_eval_order_fn, "wb") as f:
-    pickle.dump(A_eval_order,f)
-with open(A_reduced_exprs_fn, "wb") as f:
-    pickle.dump(A_reduced_exprs,f)
+    symbol_generator = (sp.Symbol(f'Ax{i}') for i in count())
+    replacements, reduced_exprs = sp.cse(A_expr_list, symbols=symbol_generator, optimizations='basic')
+    replacements, reduced_exprs = limit_cse(replacements, reduced_exprs, 3)
+    with open(A_replacements_fn, 'wb') as f:
+        pickle.dump(replacements, f)
+    with open(A_reduced_exprs_fn, 'wb') as f:
+        pickle.dump(reduced_exprs, f)
 
 
 B_expr_list = [sp.parsing.sympy_parser.parse_expr(s)*B_intcoeff for s in B_expr_str_list]
